remove message_handler file
